requirements.txt
line-bot-sdk
flask
gunicorn
opencv-python
mediapipe
imutils
pandas 
openpyxl
obspy
scikit-learn
tqdm
PyWavelets
